backend/pyproject.toml
[project]
name = "job-matching-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai-agents>=0.0.5",
    "openai>=1.50.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.9",
    "aiosmtplib>=3.0.0",
    "dnspython>=2.6.0",
    "pdfminer-six>=20221105",
    "python-docx>=1.1.0",
    "bcrypt>=4.1.0",
    "PyJWT>=2.8.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "email-validator>=2.1.0",
]
